% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{adapter_capabilities}
\alias{adapter_capabilities}
\title{Return the capability matrix of an adapter}
\usage{
adapter_capabilities(adapter)
}
\arguments{
\item{adapter}{A \code{multifer_adapter} object.}
}
\value{
A \code{multifer_capability_matrix}.
}
\description{
Return the capability matrix of an adapter
}
