% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_operator.R
\name{as_linear_operator}
\alias{as_linear_operator}
\alias{as_linear_operator.multifer_linear_operator}
\alias{as_linear_operator.matrix}
\title{Coerce an object to a \code{multifer_linear_operator}.}
\usage{
as_linear_operator(x, ...)

\method{as_linear_operator}{multifer_linear_operator}(x, ...)

\method{as_linear_operator}{matrix}(x, ...)
}
\arguments{
\item{x}{Object to coerce.}

\item{...}{Method-specific arguments.}
}
\value{
A \code{multifer_linear_operator}.
}
\description{
Coerce an object to a \code{multifer_linear_operator}.
}
