% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_result.R
\name{empty_infer_result}
\alias{empty_infer_result}
\title{An empty \code{infer_result}}
\usage{
empty_infer_result()
}
\value{
An \code{infer_result} with zero units.
}
\description{
Convenience constructor for tests and stub adapters. Returns an
\code{infer_result} with zero units and empty tables, but with valid
(zero-length) shapes everywhere.
}
