% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{exchangeable_rows}
\alias{exchangeable_rows}
\title{Exchangeable rows}
\usage{
exchangeable_rows()
}
\value{
A \code{multifer_design} object.
}
\description{
Rows are iid-exchangeable. The default design for one-block
decompositions without grouping structure.
}
