% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capability_matrix.R
\name{is_capability_matrix}
\alias{is_capability_matrix}
\title{Test whether an object is a capability matrix}
\usage{
is_capability_matrix(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Test whether an object is a capability matrix
}
