% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_role_schema.R
\name{is_data_role}
\alias{is_data_role}
\title{Test whether an object is a data role}
\usage{
is_data_role(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a data role
}
