% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{is_geneig_operator}
\alias{is_geneig_operator}
\title{Test whether an object is a validated geneig operator pair}
\usage{
is_geneig_operator(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Test whether an object is a validated geneig operator pair
}
