% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{is_geometry}
\alias{is_geometry}
\title{Test whether an object is a multifer geometry}
\usage{
is_geometry(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Test whether an object is a multifer geometry
}
