% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{is_infer_adapter}
\alias{is_infer_adapter}
\title{Test whether an object is a multifer adapter}
\usage{
is_infer_adapter(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a multifer adapter
}
