% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_problem.R
\name{is_infer_problem}
\alias{is_infer_problem}
\title{Test whether an object is a compiled inferential problem}
\usage{
is_infer_problem(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a compiled inferential problem
}
