% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_recipe.R
\name{is_infer_recipe}
\alias{is_infer_recipe}
\title{Test whether an object is a compiled infer recipe}
\usage{
is_infer_recipe(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a compiled infer recipe
}
