% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_result.R
\name{is_infer_result}
\alias{is_infer_result}
\title{Test whether an object is an \code{infer_result}}
\usage{
is_infer_result(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Test whether an object is an \code{infer_result}
}
