% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_operator.R
\name{is_linear_operator}
\alias{is_linear_operator}
\title{Test whether \code{x} is a \code{multifer_linear_operator}.}
\usage{
is_linear_operator(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether \code{x} is a \code{multifer_linear_operator}.
}
