% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_role_schema.R
\name{is_null_spec}
\alias{is_null_spec}
\title{Test whether an object is a null specification}
\usage{
is_null_spec(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a null specification
}
