% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\name{is_relation}
\alias{is_relation}
\title{Test whether an object is a multifer relation}
\usage{
is_relation(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Test whether an object is a multifer relation
}
