% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_shape.R
\name{is_typed_shape}
\alias{is_typed_shape}
\title{Test whether an object is a multifer typed shape}
\usage{
is_typed_shape(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Test whether an object is a multifer typed shape
}
