% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{list_infer_adapters}
\alias{list_infer_adapters}
\title{List all registered adapter ids}
\usage{
list_infer_adapters()
}
\value{
Character vector of registered adapter ids (possibly empty).
}
\description{
List all registered adapter ids
}
