% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_operator.R
\name{lo_apply}
\alias{lo_apply}
\title{Apply a linear operator to a matrix.}
\usage{
lo_apply(op, v)
}
\arguments{
\item{op}{A \code{multifer_linear_operator}.}

\item{v}{A numeric matrix or vector compatible with \code{op}.}
}
\value{
A numeric matrix.
}
\description{
Apply a linear operator to a matrix.
}
