% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_operator.R
\name{lo_apply_t}
\alias{lo_apply_t}
\title{Apply the transposed linear operator.}
\usage{
lo_apply_t(op, v)
}
\arguments{
\item{op}{A \code{multifer_linear_operator}.}

\item{v}{A numeric matrix or vector.}
}
\value{
A numeric matrix.
}
\description{
Apply the transposed linear operator.
}
