% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_svd_cache.R
\name{svd_cache_hits}
\alias{svd_cache_hits}
\title{Number of cache hits.}
\usage{
svd_cache_hits(cache = NULL)
}
\arguments{
\item{cache}{Cache environment. \code{NULL} uses the package default.}
}
\value{
Integer.
}
\description{
Number of cache hits.
}
