% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{unregister_infer_adapter}
\alias{unregister_infer_adapter}
\title{Remove a registered adapter}
\usage{
unregister_infer_adapter(adapter_id)
}
\arguments{
\item{adapter_id}{Character scalar.  Errors if not registered.}
}
\value{
Invisibly returns \code{adapter_id}.
}
\description{
Remove a registered adapter
}
